:heavy_check_mark: removed deprecated pytorch files, start rewrite
