--- index.html
@@ -1,22 +1,0 @@
-<html>
-<body>
-<center>
-    <br><br><br>
-    <img src="./images/stonkfish.png">
-    <br><br><br>
-    <big>
-        DO YOU WISH YOU WERE A GRANDMASTER? OR THAT YOU COULD BEAT STOCKFISH?<br>
-        WELL NOW YOU CAN, WITH THIS ONE SIMPLE TRICK!
-    </big>
-    <br><br><br>
-    <br><br><br>
-    <br><br><br>
-    <img src="./images/birch.png"> <br>
-    <big>
-        PRESENTING - BIRCHNET <br>
-        A STATIC EVALUATOR FOR CHESS POSITIONS SO THAT YOU(!) CAN BECOME THE NEXT MAGNUS CARLSEN!
-    </big>
-    <br><br><br>
-</center>
-</body>
-</html>
