+++ src/pychess_cli.py
@@ -1,0 +1,6 @@
+class PychessCLI:
+	def __init__(self):
+		self.mode = 'cli'
+
+	def start(self):
+		print("[*]  starting pychess instance with mode {}".format(self.mode))
