+++ src/pychess_gui.py
@@ -1,0 +1,6 @@
+class PychessGUI:
+	def __init__(self):
+		self.mode = 'gui'
+
+	def start(self):
+		print("[*]  starting pychess instace with mode {}".format(self.mode))
